travel_agent/modules/web_crawler.py
import requests
from bs4 import BeautifulSoup
import time
import random
from ..config.settings import USER_AGENTS, REQUEST_INTERVAL

class WebCrawler:
    def get_attractions(self, city: str, max_attractions=5):
        """Searches for top attractions in a city."""
        print(f"Searching for attractions in {city}...")
        try:
            # This is a placeholder implementation. In a real application,
            # you would use a more robust web scraping solution or an API.
            attractions = []
            for i in range(max_attractions):
                attractions.append({
                    "name": f"Attraction {i+1} in {city}",
                    "description": f"Description for Attraction {i+1}",
                    "image_url": self.get_image_url(f"Attraction {i+1}")
                })
                time.sleep(REQUEST_INTERVAL)
            return attractions
        except Exception as e:
            print(f"Error while searching for attractions: {e}")
            return []

    def get_image_url(self, query: str):
        """Searches for an image URL for a given query."""
        print(f"Searching for image: {query}")
        try:
            # This is a placeholder for a real image search
            return f"https://via.placeholder.com/300x200.png?text={query.replace(' ', '+')}"
        except Exception as e:
            print(f"Error while searching for image: {e}")
            return ""