requirements.txt
pypinyin
requests
google-adk
beautifulsoup4
lxml
Jinja2
google-generativeai
google-cloud-aiplatform
fastapi
uvicorn
python-dotenv
pypinyin
httpx
httpx-sse
tenacity
crawl4ai
markdownify
python-dateutil
aiohttp
asyncio
pytz
pillow
selenium
webdriver_manager